refactor: extract edge insertion composables
