+++ ui/src/lib/components/FlowCanvasInner.svelte
@@ -60,18 +60,7 @@
   } from '../../stores/canvas.store';
 
   import { getObjectNameFromExpr } from '$lib/objects/object-definitions';
-  import { objectSchemas } from '$lib/objects/schemas';
-  import {
-    createEdgeInsertionPreview,
-    getCenteredNodeInsertionPosition,
-    getEdgeInsertionPosition,
-    planEdgeInsertion
-  } from '$lib/canvas/edge-insertion';
-  import {
-    applyEdgeInsertionPipePreset,
-    getEdgeInsertionObjectName,
-    prepareNodeForEdgeInsertion
-  } from '$lib/canvas/edge-insertion-adapters';
+  import { useEdgeInsertion } from '$lib/canvas/use-edge-insertion.svelte';
   import { deleteSearchParam } from '$lib/utils/search-params';
 
   import { Toaster } from '$lib/components/ui/sonner';
@@ -133,8 +122,6 @@
 
   import {
     HistoryManager,
-    AddNodeCommand,
-    AddEdgesCommand,
     DeleteNodesCommand,
     ReplaceNodesCommand,
     UpdateNodeDataCommand,
@@ -167,15 +154,6 @@
   import { buildChatViewportSummary } from '$lib/ai/chat/viewport-summary';
 
   const AUTOSAVE_INTERVAL = 2500;
-  // Svelte Flow raises a selected edge to z-index 1000. Keep the inserted node
-  // above that layer so the selected wire is visually split by the node body.
-  const INSERTED_NODE_Z_INDEX = 1001;
-
-  interface PendingEdgeInsertion {
-    edge: Edge;
-    previewEdgeIds: string[];
-  }
-
   // Initial nodes and edges
   let nodes = $state.raw<Node[]>([]);
   let edges = $state.raw<Edge[]>([]);
@@ -203,6 +181,7 @@
 
   // Node operations service for creating/deleting/replacing nodes
   const nodeOps = new NodeOperationsService(canvasContext);
+  const edgeInsertion = useEdgeInsertion(canvasContext, nodeOps);
 
   // AI operations service for AI-related node insertion/editing
   const aiOps = new AiOperationsService(canvasContext, nodeOps);
@@ -408,8 +387,6 @@
 
   let selectedNodeIds = $state.raw<string[]>([]);
   let selectedEdgeIds = $state.raw<string[]>([]);
-  let pendingEdgeInsertion = $state.raw<PendingEdgeInsertion | null>(null);
-  const confirmedQuickInsertNodeIds = new SvelteSet<string>();
   let visualGroupSelectionStartIds = $state.raw<string[]>([]);
 
   // Track node positions at drag start for undo/redo
@@ -894,11 +871,8 @@
       saveAs: () => (showSavePatchModal = true),
       triggerAiPrompt,
       checkGeminiApiKey: checkAndHandleGeminiApiKey,
-      quickAddNode: () => {
-        const edge = getSingleSelectedEdge();
-        const position = edge ? getEdgeInsertionPosition(edge, nodes) : null;
-        createQuickInsertNode(position ?? screenToFlowPosition(lastMousePosition), edge);
-      },
+      quickAddNode: () =>
+        edgeInsertion.quickAdd(selectedEdgeIds, screenToFlowPosition(lastMousePosition)),
       toggleAllPreviews: () => {
         const willDisable = !$allPreviewsDisabled;
         $allPreviewsDisabled = willDisable;
@@ -1127,33 +1101,14 @@
     }
   }
 
-  // Handle Quick Add confirmation - record the final node to history
-  async function handleQuickAddConfirmed(event: {
+  const handleQuickAddConfirmed = (event: {
     type: 'quickAddConfirmed';
     finalNodeId: string;
     objectName: string;
-  }) {
-    if (confirmedQuickInsertNodeIds.has(event.finalNodeId)) return;
+  }) => edgeInsertion.confirmQuickAdd(event.finalNodeId, event.objectName);
 
-    confirmedQuickInsertNodeIds.add(event.finalNodeId);
-    await recordInsertedNode(event.finalNodeId, event.objectName);
-  }
-
-  // Handle Quick Add cancellation - remove node directly without history
-  function handleQuickAddCancelled(event: { type: 'quickAddCancelled'; nodeId: string }) {
-    // Remove the node directly, bypassing SvelteFlow's onbeforedelete (which records to history)
-    nodes = nodes.filter((n) => n.id !== event.nodeId);
-
-    const pending = pendingEdgeInsertion;
-    pendingEdgeInsertion = null;
-
-    if (!pending) return;
-
-    edges = [
-      ...edges.filter((edge) => !pending.previewEdgeIds.includes(edge.id)),
-      ...(edges.some((edge) => edge.id === pending.edge.id) ? [] : [pending.edge])
-    ];
-  }
+  const handleQuickAddCancelled = (event: { type: 'quickAddCancelled'; nodeId: string }) =>
+    edgeInsertion.cancelQuickAdd(event.nodeId);
 
   // Handle ObjectNode data commit (undo tracking for expr/name/params changes)
   function handleObjectDataCommit(event: ObjectDataCommitEvent) {
@@ -1213,162 +1168,15 @@
     const viewportCenterX = window.innerWidth / 2;
     const viewportCenterY = window.innerHeight / 2;
 
-    // Convert to flow coordinates (accounts for pan and zoom)
-    const edge = pendingEdgeInsertion?.edge;
-    const position =
-      (edge && getEdgeInsertionPosition(edge, nodes)) ??
-      screenToFlowPosition({ x: viewportCenterX, y: viewportCenterY });
-    const nodeId = nodeOps.createNodeFromName(
-      edge ? getEdgeInsertionObjectName(name) : name,
-      position,
-      { skipHistory: true }
+    await edgeInsertion.selectObject(
+      name,
+      screenToFlowPosition({ x: viewportCenterX, y: viewportCenterY })
     );
-    await recordInsertedNode(nodeId);
-  }
-
-  function getSingleSelectedEdge(): Edge | undefined {
-    if (selectedEdgeIds.length !== 1) return undefined;
-    return edges.find((edge) => edge.id === selectedEdgeIds[0]);
   }
 
   function openObjectBrowser() {
-    const edge = getSingleSelectedEdge();
-    pendingEdgeInsertion = edge ? { edge, previewEdgeIds: [] } : null;
+    edgeInsertion.beginObjectBrowser(selectedEdgeIds);
     $isObjectBrowserOpen = true;
-  }
-
-  function createQuickInsertNode(position: { x: number; y: number }, edge?: Edge) {
-    const nodeId = nodeOps.createNode('object', position, undefined, { skipHistory: true });
-    if (!edge) return;
-
-    const previewEdges = createEdgeInsertionPreview(edge, nodeId, [
-      canvasContext.nextEdgeId(),
-      canvasContext.nextEdgeId()
-    ]);
-
-    pendingEdgeInsertion = { edge, previewEdgeIds: previewEdges.map((preview) => preview.id) };
-    nodes = nodes.map((node) =>
-      node.id === nodeId ? { ...node, zIndex: INSERTED_NODE_Z_INDEX } : node
-    );
-    edges = [...edges.filter((candidate) => candidate.id !== edge.id), ...previewEdges];
-    void centerQuickInsertPreview(nodeId, edge);
-  }
-
-  async function centerQuickInsertPreview(nodeId: string, edge: Edge) {
-    await tick();
-    await new Promise<void>((resolve) => requestAnimationFrame(() => resolve()));
-
-    const node = nodes.find((candidate) => candidate.id === nodeId);
-    if (!node || pendingEdgeInsertion?.edge.id !== edge.id) return;
-
-    const position = getCenteredNodeInsertionPosition(edge, nodes, node);
-    if (!position) return;
-
-    nodes = nodes.map((candidate) =>
-      candidate.id === nodeId ? { ...candidate, position } : candidate
-    );
-  }
-
-  function getNodeObjectName(node: Node): string | undefined {
-    if (node.type === 'object') {
-      const data = node.data as { name?: string; expr?: string };
-      return data.name ?? (data.expr ? getObjectNameFromExpr(data.expr) : undefined);
-    }
-
-    return node.type;
-  }
-
-  async function recordInsertedNode(nodeId: string, quickInsertObjectName?: string) {
-    await tick();
-    await new Promise<void>((resolve) => requestAnimationFrame(() => resolve()));
-
-    let node = nodes.find((candidate) => candidate.id === nodeId);
-    if (!node) return;
-
-    const pending = pendingEdgeInsertion;
-    const edge = pending?.edge;
-    pendingEdgeInsertion = null;
-
-    if (!edge) {
-      historyManager.record(new AddNodeCommand({ ...node }, canvasAccessors));
-      return;
-    }
-
-    if (quickInsertObjectName) {
-      node = applyEdgeInsertionPipePreset(node, quickInsertObjectName);
-    }
-
-    node = prepareNodeForEdgeInsertion(node, edge);
-
-    const centeredPosition = getCenteredNodeInsertionPosition(edge, nodes, node);
-
-    if (centeredPosition) {
-      node = { ...node, position: centeredPosition };
-    }
-
-    node = { ...node, zIndex: INSERTED_NODE_Z_INDEX };
-    nodes = nodes.map((candidate) => (candidate.id === nodeId ? node : candidate));
-
-    const plan = planEdgeInsertion(
-      edge,
-      node,
-      nodes.find((candidate) => candidate.id === edge.target),
-      objectSchemas,
-      getNodeObjectName
-    );
-
-    if (!plan) {
-      if (!edges.some((candidate) => candidate.id === edge.id)) {
-        edges = [
-          ...edges.filter((candidate) => !pending.previewEdgeIds.includes(candidate.id)),
-          edge
-        ];
-      } else if (pending.previewEdgeIds.length > 0) {
-        edges = edges.filter((candidate) => !pending.previewEdgeIds.includes(candidate.id));
-      }
-
-      historyManager.record(new AddNodeCommand({ ...node }, canvasAccessors));
-      return;
-    }
-
-    const insertedEdges: Edge[] = [
-      {
-        id: canvasContext.nextEdgeId(),
-        source: edge.source,
-        sourceHandle: plan.sourceHandle,
-        target: node.id,
-        targetHandle: plan.insertedInletHandle,
-        zIndex: 0
-      },
-      {
-        id: canvasContext.nextEdgeId(),
-        source: node.id,
-        sourceHandle: plan.insertedOutletHandle,
-        target: edge.target,
-        targetHandle: plan.targetHandle,
-        zIndex: 0
-      }
-    ];
-
-    // The new node is already mounted (so Quick Add can edit it); record the equivalent
-    // atomic command for undo/redo, then apply only the edge replacement now.
-    edges = [
-      ...edges.filter(
-        (candidate) => candidate.id !== edge.id && !pending.previewEdgeIds.includes(candidate.id)
-      ),
-      ...insertedEdges
-    ];
-
-    historyManager.record(
-      new BatchCommand(
-        [
-          new AddNodeCommand({ ...node }, canvasAccessors),
-          new DeleteEdgesCommand([edge], canvasAccessors),
-          new AddEdgesCommand(insertedEdges, canvasAccessors)
-        ],
-        `Insert ${getNodeObjectName(node) ?? 'object'} into connection`
-      )
-    );
   }
 
   const isValidConnection: IsValidConnection = (connection) => {
@@ -1404,15 +1212,13 @@
   }
 
   function insertObjectWithButton() {
-    const edge = getSingleSelectedEdge();
-    const edgePosition = edge ? getEdgeInsertionPosition(edge, nodes) : null;
     const position = screenToFlowPosition({
       x: $isMobile ? window.innerWidth / 2 : window.innerWidth / 2 - 200,
       y: $isMobile ? window.innerHeight / 3 : 50
     });
 
     setTimeout(() => {
-      createQuickInsertNode(edgePosition ?? position, edge);
+      edgeInsertion.quickAdd(selectedEdgeIds, position);
     }, 50);
   }
 
@@ -1910,7 +1716,7 @@
     <ObjectBrowserModal
       bind:open={$isObjectBrowserOpen}
       onSelectObject={handleObjectBrowserSelect}
-      onClose={() => (pendingEdgeInsertion = null)}
+      onClose={edgeInsertion.clearPendingInsertion}
     />
 
     <!-- Settings Modal -->
